Immediate allocation of required resources and small additional fixes (#8)
This change contains the improvement of the nodes allocation immediate after getting the request & job in query.

The idea is to move to alter the logic of provisioning by looking at the actual Queue demand for the requested label to provision no more than not assigned waiting in queue jobs and btw to move to CloudState which was introduced in 2.259 . If we will not limit the amount of required nodes - jenkins will request tons of them (~2 times more than needed). The not used agents will be destroyed in 5 mins, but for limited budget it's not that great anyway.

Also fixed pom.xml which contained CR as in windows and missed `when` field documentation of `auqariumSnapshot` step.