Added channel close listener and interruption of the pipelines in case Aquarium was the cause. (#26)
With this change on any channel close plugin will check if the cause is Aquarium (timeout or any other issue with application state) and will notify pipeline in this case. Otherwise it will just wait because the closed channel could mean that agent temporarily disconnected.

Also the change contains some cleanups and small fixes.