Removed action on disconnected agent for now to not kill the entire pipeline (#29)
During real pipeline testing after deployment found there is a logical error in the listener and action actually sends abort to the entire pipeline rather then just one thread executed on computer. Will need to rethink it further in #18, but for now disabled the action so only log is here.