Hotfix for the issue with client timeouts - 10s is too quick (#33)
By default timeout is set to 10s and when fish node is busy - it triggers a cycle of death, where the created nodes becomes old and obsolete - because their instances already destroyed but thay can't be deleted due to the timeout of the request. 5 mins is a safety precaution, until we will find a better way.